logctx trace can be string-adjacent
If we put a trace ID into the context, but it's not an actual string
(Stringer, subtype of string), we wouldn't fetch it out.

Instead, if the key is present and is string-like, use it.
If it's present and not string-like, add `!BADVALUE-` and use it.
It's very confusing if the key is ignored entirely when set.